Well, I thought that dates had been working, and they were, sort of, but now the Asset dates work much better. They are now using the to/fromDBDate() functions for more universal storage.
As well, the Time class has been fixed up to allow for the passing of many parameters, from date-strings of several formats, to a DateTime object.

The Asset Content still isn't working with the new DataManager. It seems like the records are saving, but when trying to get the content, no Record objects can be retrieved from the RecordSet.