Now the _loadAgents method greats NEW properties arrays to pass to the agents without their references getting all mixed up.
Jesus honkey that was hard to get to happen. First, resetting the array for each Agent wiped out the previous agent's array because they were all pointing to the same thing. Then, each new Property was really just a reference to the others. Ahhhhhh. Oh well. it works now if it isn't ellegant.