Current implementation:  - The DR Manager uses a particular Hierarchy.  - Each of the Root-nodes is a DR, all children are assets.
Problem:
 - Since DRs are defined by being root nodes, there is way to create
parent nodes (AZ qualifiers) which could contain all DR's as a unit.
 - without the ability to create qualifiers that would be ancestors of
all DRs, there is no way to create global AZ's that cascade to all the
DR's.
 - This could be done in application code, but would be messy and defeat
the purpose of using the AZ system.

Fix:
    Since DR doesn't have a type itself, a node having a predefined type
such as DR::Harmoni::DR could be the definition of a DR.  Any
descendants of that node would be its Assets. Any parents or siblings of
that node would be ignored by the DR Manager.

This fix has been implemented and seems to be working.