Adding in a pass by reference to fix an Chronology bug.
The _normalize() method needed an array passed by reference to act on.
In the PHP 5 migration the reference got deleted by accident, so the
normalization wasn't being done. This had the effect of not rolling over
days when adding durations that bumped a date and time over the length of
a day.