Harmoni now allows configuration of actions that allow the SESSION ID in URLs
if we are using an SID for an action configured to allow it in the url,
the Set-Cookie header is overwritted with a bogus one to prevent session
fixation.

This header line will not affect any existing cookies in the browser, it
will simply prevent the Set-Cookie header from being sent with a valid
session id. In this way, if a person sends another to the a URL where
we are allowing the session id in the URL, that person's browser will
only use the session id in the URL for this one request, not all future
ones.