Admin-Act-As-User authentication now is not reliant on usernames.
The Admin-Act-As-User now has its own token collector that performs a
search for users and allows selection from a list. This is dependent
on the existence of an auth/search_users action (provided in Polyphony).
Using this feature also requires the AuthenticationManager to be configured
with 'jquery_src', 'jquery_autocomplete_src', and 'jquery_autocomplete_css'
parameters.